feat: decimal consumable quantities (#98 discussion)
Half a can of spray is 0.5 (Discussion #98): stock, consumption links,
restock deltas and reorder thresholds accept decimals end to end, so the
CAN stays the unit (price 4.99/can) instead of a millilitre workaround
that broke on price precision.

Unit-dimension sweep (the #58/#59 audit checklist — every consumer):
- helpers/parts: round_qty() = canonical 2-decimal rounding, whole numbers
  collapse to int; _clean_stock/floats; sanitize_consumes_parts keeps 0.5,
  zero/invalid still falls back to 1; part_is_low compares floats
- storage: set/get_part_stock float, rounded at the single write choke point
- parts_runtime: float deltas/absolutes/quantities in the consume + restock
  + buy-task math
- WS schemas: part stock/restock_quantity/delta and the complete-dialog
  quantity-bought accept numbers (Coerce(float) with ranges)
- io import: float stock survives the JSON round-trip
- frontend: parseFloat + step=0.01 on restock input, part form (Number()
  already float-capable), complete-dialog quantity fields, task-dialog
  consumes-quantity input

Tests: #98 scenario pinned (0.5 consumption round-trip, decimal stock,
fractional low-threshold comparison, rounding collapse); full backend
2,775 @ gate; frontend suite green; ruff + mypy clean.

+++ custom_components/maintenance_supporter/coordinator.py
@@ -964,7 +964,7 @@ class MaintenanceCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         completed_by: str | None = None,
         photo_doc_id: str | None = None,
         reading_value: float | None = None,
-        restock_quantity: int | None = None,
+        restock_quantity: float | None = None,
     ) -> None:
         """Mark a task as completed and persist."""
         merged = self._get_merged_tasks_data()
